Refactored code so we don't have to use an external library. Now files in root path of Downloads and inside subfolders supported.
