--- app/src/main/java/swati4star/createpdf/pickit/CallBackTask.java
@@ -1,7 +1,0 @@
-package swati4star.createpdf.pickit;
-
-interface CallBackTask {
-    void PickiTonPreExecute();
-    void PickiTonProgressUpdate(int progress);
-    void PickiTonPostExecute(String path, boolean wasDriveFile, boolean wasSuccessful, String reason);
-}
