--- app/src/main/java/swati4star/createpdf/pickit/PickiTCallbacks.java
@@ -1,7 +1,0 @@
-package swati4star.createpdf.pickit;
-
-public interface PickiTCallbacks {
-    void PickiTonStartListener();
-    void PickiTonProgressUpdate(int progress);
-    void PickiTonCompleteListener(String path, boolean wasDriveFile, boolean wasUnknownProvider, boolean wasSuccessful, String Reason);
-}
